Have all files, just need to add anilist videos now

+++ jikan_get.py
@@ -14,20 +14,25 @@ import re
 # methods
 # -------
 
-def test():
-    tokenized = 'https://myanimelist.net/people/142/Takeshi_Aono'.split('/')
-    print(tokenized)
+# Counts the number of files in the given folder
+def count_files(folder):
+    print len([name for name in os.listdir(folder) if os.path.isfile(os.path.join(folder, name))])
 
-
+# Used to get anime, starting at id 1 and going until you have num_items
 def jikanget(url_type, num_items, folder_name, filter_porn=True):
 
     id = 1
-    count = 1
+    if os.path.exists(folder_name) :
+        count = len([name for name in os.listdir(folder_name) if os.path.isfile(os.path.join(folder_name, name))])
+    else :
+        count = 0;
 
     while count <= num_items:
         url = url_type + str(id) + '/characters_staff'
         r = requests.get(url)
         print(r.status_code)
+        print(count)
+        add_outcome = 'FAILED'
 
         if r.status_code != 404:
             r = r.json()
@@ -42,22 +47,30 @@ def jikanget(url_type, num_items, folder_name, filter_porn=True):
             
             # only add the anime if it's not porn, will always run if filter_porn
             # is set to false
-            if has_porn != True :
-	            file_name = folder_name + '/' + str(id)
-	            print(file_name)
-	            if not os.path.exists(os.path.dirname(file_name)):
-	                try:
-	                    os.makedirs(os.path.dirname(file_name))
-	                except OSError as exc:  # Guard against race condition
-	                    if exc.errno != errno.EEXIST:
-	                        raise
+            if has_porn != True:
+                file_name = folder_name + '/' + str(id)
+                print file_name
+                if not os.path.exists(os.path.dirname(file_name)):
+                    try:
+                        os.makedirs(os.path.dirname(file_name))
+                    except OSError, exc:
 
-	            with open(str(file_name), 'w') as outfile:
-	                json.dump(r, outfile)
-	            count += 1
+                                # Guard against race condition
+
+                        if exc.errno != errno.EEXIST:
+                            raise
+                if os.path.isfile(file_name) == False:
+                    with open(str(file_name), 'w') as outfile:
+                        json.dump(r, outfile)
+                        add_outcome = 'SUCCESS'
+                    count += 1
 
         id += 1
+        print(add_outcome)
+        print('')
 
+# Used to get each anime's matching manga, used to check and compare the title attribute of each one.
+# Was not used for characters or people because this only checks top level dict keys
 def get_matches(read_match, write_match, read_folder, write_folder, url_type, *additional_data):
 
     id = 220
@@ -111,6 +124,8 @@ def get_matches(read_match, write_match, read_folder, write_folder, url_type, *a
         print(add_outcome)
         print('')
 
+# Runs through the anime we have locally, gets the first main character, gets the Japanese voice actor for that character
+# Limited to one per show
 def get_matches_people(read_folder, write_folder, url_type):
 
     count = 0
@@ -128,7 +143,6 @@ def get_matches_people(read_folder, write_folder, url_type):
                                     if actor['role'] == 'Japanese' and found != True:
                                         if 'url' in actor:
                                             tokenized_id = (actor['url'].split('/'))[4]
-                                            print(tokenized_id)
 
                                             url = url_type + tokenized_id
                                             r = requests.get(url).json()
@@ -148,73 +162,45 @@ def get_matches_people(read_folder, write_folder, url_type):
                                                 found = True
                                             count += 1
                                             print('Count: ' + str(count))
+                                            print('')
 
 
-def get_matches_character(read_match, write_match, read_folder, write_folder, url_type, *additional_data):
+# Runs through the anime we have locally, gets the first main character, and adds them to jikan_character
+# Limited to one per show
+def get_matches_character(read_folder, write_folder, url_type):
 
-    id = 127
-    count = 56
-    num_items = 100
+    count = 0
 
-    while count <= num_items:
-        print('here 1')
-        # get a json page based off of id
-        url = url_type + str(id) + additional_data
-        print('here 2')
-        r = requests.get(url)
-        print(r.status_code)
-        print(id)
-        print(count)
-        add_outcome = 'FAILED'
+    for filename in os.listdir(read_folder):
+        found = False
+        with open(read_folder + '/' + filename) as datafile:
+            data = json.load(datafile)
+            if 'character' in data:
+                for character in data['character']:
+                    if 'role' in character:
+                        if character['role'] == 'Main' and found != True:
+                            if 'url' in character:
+                                tokenized_id = (character['url'].split('/'))[4]
 
-        # if the daily limit is not exceeded
-        if r.status_code == 429:
-            raise StopIteration
+                                url = url_type + tokenized_id
+                                r = requests.get(url).json()
 
-        # if the page returns something
-        if r.status_code == 200:
-            # turn the json page into a dict
-            r = r.json()
-            if r != False:
-                # make sure our match field exists
-                if write_match in r:
-                    match_data = r[write_match]
-                    match_data = re.findall(r"[\w']+", match_data)
-                    match_set = set(match_data)
-                    print(match_set)
-                    # check to see if that item's match field matches any of the titles in anime
-                    # runs through all the files in the give folder
-                    for readfile in os.listdir(read_folder):
-                        with open(read_folder + '/' + readfile) as datafile:
-                            data = json.load(datafile)
-                            for character in data['character']:
-                                if 'name' in character:
-                                    char_name = re.findall(r"[\w']+", character['name'])
-                                    name_set = set(char_name)
-                                    if name_set == match_set:
-                                        # if the field matches between file
-                                            # name the file after it's id
-                                        writefile = write_folder + '/' + str(id)
-                                        # create the folder if it does not exist in this dir
-                                        if not os.path.exists(os.path.dirname(writefile)):
-                                            try:
-                                                os.makedirs(os.path.dirname(writefile))
-                                            except OSError as exc:  # Guard against race condition
-                                                if exc.errno != errno.EEXIST:
-                                                    raise
-                                        duplicate = False            
-                                        for file in os.listdir(write_folder):
-                                            if file == str(id):
-                                                duplicate = True
-                                        # write the file
-                                        if duplicate != True:
-                                            with open(writefile, 'w') as outfile:
-                                                json.dump(r, outfile)
-                                                add_outcome = 'SUCCESS'
-                                            count += 1
-        id += 1
-        print(add_outcome)
-        print('')
+                                file_name = write_folder + '/' + tokenized_id
+                                # create the folder if it does not exist in this dir
+                                if not os.path.exists(os.path.dirname(file_name)):
+                                    try:
+                                        os.makedirs(os.path.dirname(file_name))
+                                    except OSError as exc:  # Guard against race condition
+                                        if exc.errno != errno.EEXIST:
+                                            raise
+                                # write the file
+                                with open(file_name, 'w') as outfile:
+                                    json.dump(r, outfile)
+                                    print(data['title'])
+                                    found = True
+                                count += 1
+                                print('Count: ' + str(count))
+                                print('')
 
 
 # ----
@@ -222,10 +208,11 @@ def get_matches_character(read_match, write_match, read_folder, write_folder, ur
 # ----
 
 if __name__ == "__main__":
-    # titles used to match anime and manga
-    # voice actor page links used to match anime and people
-    # test()
+
+    # count_files('jikan_anime')
+
     # jikanget('http://jikan.me/api/anime/', 100, 'jikan_anime', True)
+    # jikanget('http://jikan.me/api/manga/', 100, 'jikan_manga', True)
     # get_matches('title', 'title', 'jikan_anime', 'jikan_manga', 'http://jikan.me/api/manga/', '/characters_staff')
-    get_matches_people("jikan_anime", "jikan_person", 'http://jikan.me/api/person/')    
-    # get_matches_character('character', 'name', 'jikan_anime', 'jikan_character', 'http://jikan.me/api/character/')
+    # get_matches_people("jikan_anime", "jikan_person", 'http://jikan.me/api/person/')    
+    # get_matches_character('jikan_anime', 'jikan_character', 'http://jikan.me/api/character/')
